requirements.txt
Flask==2.0.1
numpy==1.21.0
Pillow==8.2.0
tensorflow==2.13.0
gunicorn==20.1.0